refactor: Refactor, to add 2 buttons to make a default grid size, and change the amount with a prompt
simpfied the css to make rules simplier.  A default amount of coloured boxes was made and a button was added to allow a prompt to accept a value between 1-100 with a  check added in.  A function was made during intilatization to cause the gridsize to be automatically added during page refresh. The event listener was changed to change red and  then be turned off from the js as an experiment with different events I could use.